requirements.txt
boto3
ibm-watson
requests
pyyaml
pandas
flask
flask_cors